User commit

+++ src/main/java/com/example/demo/repository/StudentRepository.java
@@ -6,6 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.model.Student;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student,Integer>{
-
+public interface StudentRepository extends JpaRepository<Student, Long> {
 }
